+++ src/core/EdFi.DataManagementService.Core/ApiService.cs
@@ -50,6 +50,7 @@ internal class ApiService(
                     new ProvideApiSchemaMiddleware(_apiSchemaProvider, _logger),
                     new ParsePathMiddleware(_logger),
                     new ParseBodyMiddleware(_logger),
+                    new DuplicatePropertiesMiddleware(_logger),
                     new ValidateEndpointMiddleware(_logger)
                 ]);
